:judge: :copyright: legal: Fix automatic detection of Apache-2.0 license
This commit applies the full copytext of the Apache License Version 2.0
to the repository. This allows the GitHub API to correctly identify this
repository as licensed under the Apache-2.0 and label it correctly
across the GitHub user interface. Additionally, it also ensures the
complete distribution of the full license terms with digital copies
cloned from a git repository.

It also makes some of my job easier. :wink:

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>